fix(Reservation): id필드의 타입을 Long으로 변경

+++ src/main/java/roomescape/domain/Reservation.java
@@ -4,19 +4,19 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 
 public class Reservation {
-    private int id;
+    private Long id;
     private String name;
     private LocalDate reservationDate;
     private LocalTime reservationTime;
 
-    public Reservation(int id, String name, LocalDate reservationDate, LocalTime reservationTime) {
+    public Reservation(Long id, String name, LocalDate reservationDate, LocalTime reservationTime) {
         this.id = id;
         this.name = name;
         this.reservationDate = reservationDate;
         this.reservationTime = reservationTime;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
